댓글 수정, 삭제, 하나 조회 기능 추가 & md파일 수정 - 본인 댓글 아닌 경우: 401 에러 - 없는 댓글일 경우: 404 에러 - content없을 경우: 403 에러

+++ routes/comment.js
@@ -13,14 +13,14 @@ let result = { // response form
  * - parameter user_id : 로그인 한 회원 uuid
  * - parameter friend_id : 친구 추가할 회원 uuid
  */
-router.get('/:board_id', async(req, res, next)=>{
-    const user_uid = req.user.user_uid;
+router.get('/selectall/:board_id', async(req, res, next)=>{
     const board_id = req.params.board_id;
     console.log(`해당 게시물의 전체 댓글 조회 요청 = ${board_id}`);
+    result.data = '';
     try{
         await Comment.findAll({
             where:{fk_board_id: board_id},
-            include: [{
+            include: [{ // 댓글 작성자의 nickname, portrait 정보 가져오기
                 model: User,
                 attributes: ['nickname', 'portrait'],
             }]
@@ -45,6 +45,42 @@ router.get('/:board_id', async(req, res, next)=>{
     }
 });
 
+/* get one comment
+ * - parameter user_uid : 로그인 한 회원 uuid
+ * - parameter friend_id : 친구 추가할 회원 uuid
+ */
+router.get('/selectone/:comment_id', async(req, res, next)=>{
+    const comment_id = req.params.comment_id;
+    console.log(`댓글 조회 요청 = ${comment_id}`);
+    result.data = '';
+    try{
+        await Comment.findOne({
+            where:{id: comment_id},
+            include: [{
+                model: User,
+                attributes: ['nickname', 'portrait'],
+            }]
+        }).then(comment=>{
+            if(comment){
+                result.data = comment;
+                result.success = true;
+                result.message = "댓글 조회 성공";
+                return res.status(200).json(result);
+            }else{
+                result.success = false;
+                result.message = "댓글 조회 실패: 존재하지 않는 댓글입니다.";
+                return res.status(403).json(result);
+            }
+        });
+    }catch(e){
+        result.success = false;
+        result.message = e;
+        res.status(500).json(result);
+        console.error(e);
+        return next(e);
+    }
+});
+
 /* update one comment = 전체 댓글보기가 곧 해당 게시물 보기와 같음.
  * - parameter user_id : 로그인 한 회원 uuid
  * - parameter friend_id : 친구 추가할 회원 uuid
@@ -53,7 +89,12 @@ router.post('/create', isLoggedIn, async (req, res, next)=>{
     try{
         const user_uid = req.user.user_uid;
         const {board_id, comment_content} = req.body;
-    
+        result.data = '';
+        if(!comment_content) {
+            result.success = false;
+            result.message = "댓글 생성 실패: 댓글 내용(comment_content)는 반드시 입력해야 합니다.";
+            return res.status(403).json(result);
+        }
         console.log(`새 댓글 생성 요청 들어옴 = ${board_id}, ${comment_content}`);
         // board_id값은 들어오는데 
         //{ Error: (conn=161, no: 1364, SQLState: HY000) Field 'board_id' doesn't have a default value 에러 발생
@@ -88,34 +129,46 @@ router.post('/create', isLoggedIn, async (req, res, next)=>{
  * - parameter user_id : 로그인 한 회원 uuid
  * - parameter friend_id : 친구 추가할 회원 uuid
  */
-router.put('update/:comment_id', isLoggedIn, async (req, res, next)=>{
+router.put('/update/:comment_id', isLoggedIn, async (req, res, next)=>{
     const user_uid = req.user.user_uid;
-    //const comment_id = req.params.comment_id;
-    const {comment_id, comment_content} = req.body;
-
+    const comment_id = req.params.comment_id;
+    const comment_content = req.body.comment_content;
+    result.data = '';
+    if(!comment_content) {
+        result.success = false;
+        result.message = "댓글 수정 실패: 댓글 내용(comment_content)는 반드시 입력해야 합니다.";
+        return res.status(403).json(result);
+    }
     console.log(`기존 댓글 수정 요청 들어옴 = ${comment_id} / ${comment_content}`);
 
     try{
         await Comment.findOne({
             where: {id: comment_id}
         }).then(comment=>{
-            if(comment.fk_user_uid==user_uid){ // 댓글 주인만 수정 가능
-                comment_content = comment_content;
-                
-                comment.save();
-
-                result.data = comment; //삭제된 댓글 id 반환
-                result.success = true;
-                result.message = "댓글 수정 성공";
-                console.log(`Update One User's comment`, JSON.stringify(result));
-                return res.status(200).json(result);
-
+            if(comment){
+                if(comment.fk_user_uid==user_uid){ // 댓글 주인만 수정 가능
+                    comment.comment_content = comment_content;
+                    comment.save();
+    
+                    result.data = comment; //삭제된 댓글 id 반환
+                    result.success = true;
+                    result.message = "댓글 수정 성공";
+                    console.log(`Update One User's comment`, JSON.stringify(result));
+                    return res.status(200).json(result);
+    
+                }else{
+                    result.success = false;
+                    result.message = "댓글 수정 실패: 본인의 댓글만 수정할 수 있습니다.";
+                    console.log(`댓글 수정 실패: 댓글 주인 아님`, JSON.stringify(result));
+                    return res.status(401).json(result);
+                }
             }else{
                 result.success = false;
-                result.message = "댓글 수정 실패: 본인의 댓글만 삭제할 수 있습니다.";
-                console.log(`댓글 수정 실패: 댓글 주인 아님`, JSON.stringify(result));
-                return res.status(403).json(result);
+                result.message = "댓글 수정 실패: 존재하지 않는 댓글입니다.";
+                console.log(`댓글 수정 실패: 존재하지 않는 댓글`, JSON.stringify(result));
+                return res.status(404).json(result);
             }
+            
         });
 
     }catch(e){
@@ -132,10 +185,10 @@ router.put('update/:comment_id', isLoggedIn, async (req, res, next)=>{
  * - parameter user_id : 로그인 한 회원 uuid
  * - parameter friend_id : 친구 추가할 회원 uuid
  */
-router.delete('delete/:comment_id', isLoggedIn, async (req, res, next)=>{
+router.delete('/delete/:comment_id', isLoggedIn, async (req, res, next)=>{
     const user_uid = req.user.user_uid;
-    //const comment_id = req.params.comment_id;
-    const comment_id = req.body.comment_id;
+    const comment_id = req.params.comment_id;
+    result.data = '';
 
     console.log(`기존 댓글 삭제 요청 들어옴= ${comment_id}`);
     try{
@@ -143,20 +196,27 @@ router.delete('delete/:comment_id', isLoggedIn, async (req, res, next)=>{
             where: {id: comment_id}
         }).then(comment =>{
             // 댓글 주인만 삭제 가능
-            if(comment.fk_user_uid==user_uid){
-                Comment.destroy({
-                    where: {id: comment_id}
-                });
-                result.data = comment_id; //삭제된 댓글 id 반환
-                result.success = true;
-                result.message = "댓글 삭제 성공";
-                console.log(`Delete One User's comment`, JSON.stringify(result));
-                return res.status(200).json(result);
+            if(comment){
+                if(comment.fk_user_uid==user_uid){
+                    Comment.destroy({
+                        where: {id: comment_id}
+                    });
+                    result.data = comment_id; //삭제된 댓글 id 반환
+                    result.success = true;
+                    result.message = "댓글 삭제 성공";
+                    console.log(`Delete One User's comment`, JSON.stringify(result));
+                    return res.status(200).json(result);
+                }else{
+                    result.success = false;
+                    result.message = "댓글 삭제 실패: 본인의 댓글만 삭제할 수 있습니다.";
+                    console.log(`댓글 삭제 실패: 댓글 주인 아님`, JSON.stringify(result));
+                    return res.status(401).json(result);
+                }
             }else{
                 result.success = false;
-                result.message = "댓글 삭제 실패: 본인의 댓글만 삭제할 수 있습니다.";
-                console.log(`댓글 삭제 실패: 댓글 주인 아님`, JSON.stringify(result));
-                return res.status(403).json(result);
+                result.message = "댓글 삭제 실패: 이미 삭제된 댓글입니다.";
+                console.log(`댓글 삭제 실패: 존재하지 않는 댓글`, JSON.stringify(result));
+                return res.status(404).json(result);
             }
         });        
     }catch(e){
